+++ WebStore.Domain/Entities/Base/BaseEntity.cs
@@ -1,0 +1,9 @@
+﻿using WebStore.Domain.Entities.Base.Interfaces;
+
+namespace WebStore.Domain.Entities.Base
+{
+    public abstract class BaseEntity : IBaseEntity
+    {
+        public int Id { get; set; }
+    }
+}
