+++ WebStore.Domain/Entities/Base/Interfaces/IBaseEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace WebStore.Domain.Entities.Base.Interfaces
+{
+    public interface IBaseEntity
+    {
+        int Id { get; set; }
+    }
+}
