+++ WebStore.Domain/Entities/Base/Interfaces/INamedEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace WebStore.Domain.Entityes.Base.Interfaces
+﻿namespace WebStore.Domain.Entities.Base.Interfaces
 {
     public interface INamedEntity : IBaseEntity
     {
