+++ WebStore.Domain/Entities/Base/Interfaces/IOrderedEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace WebStore.Domain.Entityes.Base.Interfaces
+﻿namespace WebStore.Domain.Entities.Base.Interfaces
 {
     public interface IOrderedEntity : IBaseEntity
     {
