+++ WebStore.Domain/Entities/Base/NamedEntity.cs
@@ -1,6 +1,6 @@
-﻿using WebStore.Domain.Entityes.Base.Interfaces;
+﻿using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.Domain.Entityes.Base
+namespace WebStore.Domain.Entities.Base
 {
     public abstract class NamedEntity : BaseEntity, INamedEntity
     {
