+++ WebStore.Domain/Entities/Section.cs
@@ -1,6 +1,7 @@
-﻿using WebStore.Domain.Entityes.Base.Interfaces;
+﻿using WebStore.Domain.Entities.Base;
+using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.Domain.Entityes.Base
+namespace WebStore.Domain.Entities
 {
     public class Section : NamedEntity, IOrderedEntity
     {
