--- WebStore.Domain/Entityes/Base/BaseEntity.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using WebStore.Domain.Entityes.Base.Interfaces;
-
-namespace WebStore.Domain.Entityes.Base
-{
-    public abstract class BaseEntity : IBaseEntity
-    {
-        public int Id { get; set; }
-    }
-}
