--- WebStore.Domain/Entityes/Base/Interfaces/IBaseEntity.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace WebStore.Domain.Entityes.Base.Interfaces
-{
-    public interface IBaseEntity
-    {
-        int Id { get; set; }
-    }
-}
