--- WebStore.Domain/Entityes/Brand.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using WebStore.Domain.Entityes.Base.Interfaces;
-
-namespace WebStore.Domain.Entityes.Base
-{
-    public class Brand : NamedEntity, IOrderedEntity
-    {
-        public int Order { get; set; }
-    }
-}
